Fix USB incomplete IN transfers
The Synopsys OTG controller appears to not automatically handle sending
a ZLP even when the DIEPTSIZ register has been configured to send an
extra packet. This could lead to the USB stack becoming unresponsive
when sending an exact multiple of 1024 bytes to the host. This was fixed
by manually sending the ZLP after the rest of the transfer completes.